requirements.txt
flask>=3.0
watchdog>=4.0
Pillow>=10.0
numpy>=1.26
